테스트통과 코드 작성
Beverage 배열을 BeverageList로 추상화
자판기의 재고 딕셔너리를 BeverageStock으로 추상화
자판기의 처음 재고 상태는 BeverageStock 초기화 상태와 같음
