+++ VendingMachine/VendingMachine/BeverageList.swift
@@ -1,0 +1,14 @@
+//
+//  BeverageList.swift
+//  VendingMachine
+//
+//  Created by moon on 2018. 6. 11..
+//  Copyright © 2018년 JK. All rights reserved.
+//
+
+import Foundation
+
+struct BeverageList: Equatable {
+
+    var beverageList: [Beverage] = [Beverage]()
+}
